chore(unitests): improved repeatability
