+++ src/api/sdk/agent/agent-models.ts
@@ -1,4 +1,4 @@
-namespace AgentManagementModels {
+export namespace AgentManagementModels {
     /**
      *
      * @export
