+++ src/api/sdk/index.ts
@@ -1,6 +1,9 @@
 // Copyright (c) Siemens AG, 2019
+export * from "./agent/agent-management";
+export * from "./agent/agent-models";
 export * from "./asset/asset-management";
 export * from "./asset/asset-models";
+export * from "./common/mindsphere-sdk";
 export * from "./common/sdk-client";
 export * from "./iot/iot-timeseries";
 export * from "./iot/iot-timeseries-models";
